Refactor: move the channel subscription logic to the utils class
