+++ java/code/src/com/suse/manager/reactor/messaging/RegistrationUtils.java
@@ -17,10 +17,15 @@ package com.suse.manager.reactor.messaging;
 
 import com.redhat.rhn.common.messaging.MessageQueue;
 import com.redhat.rhn.common.validator.ValidatorResult;
+import com.redhat.rhn.domain.channel.Channel;
 import com.redhat.rhn.domain.entitlement.Entitlement;
+import com.redhat.rhn.domain.product.SUSEProduct;
+import com.redhat.rhn.domain.product.SUSEProductChannel;
+import com.redhat.rhn.domain.product.SUSEProductFactory;
 import com.redhat.rhn.domain.server.MinionServer;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.domain.server.ServerFactory;
+import com.redhat.rhn.domain.server.ServerHistoryEvent;
 import com.redhat.rhn.domain.state.PackageState;
 import com.redhat.rhn.domain.state.PackageStates;
 import com.redhat.rhn.domain.state.ServerStateRevision;
@@ -32,9 +37,16 @@ import com.redhat.rhn.domain.user.User;
 import com.redhat.rhn.manager.action.ActionManager;
 import com.redhat.rhn.manager.system.SystemManager;
 import com.redhat.rhn.taskomatic.TaskomaticApiException;
+import com.suse.manager.reactor.utils.RhelUtils;
 import com.suse.manager.reactor.utils.ValueMap;
 import com.suse.manager.webui.controllers.StatesAPI;
 import com.suse.manager.webui.services.SaltStateGeneratorService;
+import com.suse.manager.webui.services.impl.SaltService;
+import com.suse.manager.webui.utils.salt.custom.PkgProfileUpdateSlsResult;
+import com.suse.salt.netapi.calls.modules.State;
+import com.suse.salt.netapi.calls.modules.Zypper;
+import com.suse.salt.netapi.results.CmdExecCodeAll;
+import com.suse.utils.Opt;
 import org.apache.log4j.Logger;
 
 import java.util.ArrayList;
@@ -42,9 +54,14 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static java.util.Optional.ofNullable;
 
 /**
  * Common registration logic that can be used from multiple places
@@ -123,6 +140,17 @@ public class RegistrationUtils {
         }
     }
 
+    private static void triggerHardwareRefresh(MinionServer server) {
+        try {
+            ActionManager.scheduleHardwareRefreshAction(server.getOrg(), server,
+                    new Date());
+        }
+        catch (TaskomaticApiException e) {
+            LOG.error("Could not schedule hardware refresh for system: " + server.getId());
+            throw new RuntimeException(e);
+        }
+    }
+
     /**
      * Apply activation key properties to server state
      *
@@ -175,14 +203,227 @@ public class RegistrationUtils {
         });
     }
 
-    private static void triggerHardwareRefresh(MinionServer server) {
-        try {
-            ActionManager.scheduleHardwareRefreshAction(server.getOrg(), server,
-                    new Date());
+    /**
+     * Subscribes minion to channels
+     *
+     * @param saltService the Salt service instance
+     * @param minionId the minion id
+     * @param server the minion
+     * @param grains the grains
+     * @param activationKey the activation key
+     * @param activationKeyLabel the activation key label
+     */
+    public static void subscribeMinionToChannels(SaltService saltService, String minionId, MinionServer server,
+            ValueMap grains, Optional<ActivationKey> activationKey,
+            Optional<String> activationKeyLabel) {
+
+        if (!activationKey.isPresent() && activationKeyLabel.isPresent()) {
+            LOG.warn(
+                    "Default channel(s) will NOT be subscribed to: specified Activation Key " +
+                            activationKeyLabel.get() + " is not valid for minionId " +
+                            minionId);
+            addHistoryEvent(server, "Invalid Activation Key",
+                    "Specified Activation Key " + activationKeyLabel.get() +
+                            " is not valid. Default channel(s) NOT subscribed to.");
+            return;
         }
-        catch (TaskomaticApiException e) {
-            LOG.error("Could not schedule hardware refresh for system: " + server.getId());
-            throw new RuntimeException(e);
-        }
+
+        Set<Channel> channelsToAssign = Opt.fold(
+                activationKey,
+                // No ActivationKey
+                () -> {
+                    Set<SUSEProduct> suseProducts = identifyProduct(saltService, server, grains);
+                    Map<Boolean, List<SUSEProduct>> baseAndExtProd = suseProducts.stream()
+                            .collect(Collectors.partitioningBy(SUSEProduct::isBase));
+
+                    Optional<SUSEProduct> baseProductOpt = ofNullable(baseAndExtProd.get(true))
+                            .flatMap(s -> s.stream().findFirst());
+                    List<SUSEProduct> extProducts = baseAndExtProd.get(false);
+
+                    return Opt.fold(
+                            baseProductOpt,
+                            // No ActivationKey and no base product identified
+                            () -> {
+                                LOG.warn("Server " + minionId + " has no identifiable base product" +
+                                        " and will register without base channel assignment");
+                                return Collections.emptySet();
+                            },
+                            baseProduct -> {
+                                return Stream.concat(
+                                        lookupRequiredChannelsForProduct(baseProduct),
+                                        extProducts.stream()
+                                            .flatMap(ext -> recommendedChannelsByBaseProduct(baseProduct, ext))
+                                ).collect(Collectors.toSet());
+                            }
+                    );
+                },
+                ak -> {
+                    return Opt.<Channel, Set<Channel>>fold(
+                            ofNullable(ak.getBaseChannel()),
+                            // AktivationKey without base channel (SUSE Manager Default)
+                            () -> {
+                                Set<SUSEProduct> suseProducts = identifyProduct(saltService, server, grains);
+                                Map<Boolean, List<SUSEProduct>> baseAndExtProd = suseProducts.stream()
+                                        .collect(Collectors.partitioningBy(SUSEProduct::isBase));
+
+                                Optional<SUSEProduct> baseProductOpt = ofNullable(baseAndExtProd.get(true))
+                                        .flatMap(s -> s.stream().findFirst());
+                                List<SUSEProduct> extProducts = baseAndExtProd.get(false);
+
+                                return Opt.fold(
+                                        baseProductOpt,
+                                        // ActivationKey and no base product identified
+                                        () -> {
+                                            LOG.warn("Server " + minionId + "has no identifyable base product" +
+                                                    " and will register without base channel assignment");
+                                            return Collections.emptySet();
+                                        },
+                                        baseProduct -> {
+                                            return Stream.concat(
+                                                    lookupRequiredChannelsForProduct(baseProduct),
+                                                    extProducts.stream().flatMap(
+                                                            ext -> recommendedChannelsByBaseProduct(baseProduct, ext))
+                                            ).collect(Collectors.toSet());
+                                        }
+                                );
+                            },
+                            baseChannel -> {
+                                return Opt.fold(
+                                        SUSEProductFactory.findProductByChannelLabel(baseChannel.getLabel()),
+                                        () -> {
+                                            // ActivationKey with custom channel
+                                            return Stream.concat(
+                                                    Stream.of(baseChannel),
+                                                    ak.getChannels().stream()
+                                            ).collect(Collectors.toSet());
+                                        },
+                                        baseProduct -> {
+                                            // ActivationKey with vendor or cloned vendor channel
+                                            return Stream.concat(
+                                                    lookupRequiredChannelsForProduct(baseProduct.getProduct()),
+                                                    ak.getChannels().stream()
+                                                            .filter(c ->
+                                                                    c.getParentChannel() != null &&
+                                                                    c.getParentChannel().getId()
+                                                                        .equals(baseChannel.getId())
+                                                            )
+                                            ).collect(Collectors.toSet());
+
+                                        }
+                                );
+                            }
+                    );
+                }
+        );
+
+        channelsToAssign.forEach(server::addChannel);
     }
+
+    private static Set<SUSEProduct> identifyProduct(SaltService saltService, MinionServer server, ValueMap grains) {
+        if ("suse".equalsIgnoreCase(grains.getValueAsString("os"))) {
+            Optional<List<Zypper.ProductInfo>> productList =
+                    saltService.callSync(Zypper.listProducts(false), server.getMinionId());
+            return Opt.stream(productList).flatMap(pl -> {
+                return pl.stream()
+                        .flatMap(pi -> {
+                            String osName = pi.getName().toLowerCase();
+                            String osVersion = pi.getVersion();
+                            String osArch = pi.getArch();
+                            String osRelease = pi.getRelease();
+                            Optional<SUSEProduct> suseProduct =
+                                    ofNullable(SUSEProductFactory.findSUSEProduct(osName,
+                                            osVersion, osRelease, osArch, true));
+                            if (!suseProduct.isPresent()) {
+                                LOG.warn("No product match found for: " + osName + " " +
+                                        osVersion + " " + osRelease + " " + osArch);
+                            }
+                            return Opt.stream(suseProduct);
+                        });
+            }).collect(Collectors.toSet());
+        }
+        else if ("redhat".equalsIgnoreCase(grains.getValueAsString("os")) ||
+                "centos".equalsIgnoreCase(grains.getValueAsString("os"))) {
+            Optional<Map<String, State.ApplyResult>> applyResultMap = saltService
+                    .applyState(server.getMinionId(), "packages.redhatproductinfo");
+            Optional<String> centosReleaseContent = applyResultMap.map(
+                    map -> map.get(PkgProfileUpdateSlsResult.PKG_PROFILE_CENTOS_RELEASE))
+                    .map(r -> r.getChanges(CmdExecCodeAll.class)).map(c -> c.getStdout());
+            Optional<String> rhelReleaseContent = applyResultMap.map(
+                    map -> map.get(PkgProfileUpdateSlsResult.PKG_PROFILE_REDHAT_RELEASE))
+                    .map(r -> r.getChanges(CmdExecCodeAll.class)).map(c -> c.getStdout());
+            Optional<String> whatProvidesRes = applyResultMap.map(map -> map
+                    .get(PkgProfileUpdateSlsResult.PKG_PROFILE_WHATPROVIDES_SLES_RELEASE))
+                    .map(r -> r.getChanges(CmdExecCodeAll.class)).map(c -> c.getStdout());
+
+            Optional<RhelUtils.RhelProduct> rhelProduct = RhelUtils.detectRhelProduct(
+                    server, whatProvidesRes, rhelReleaseContent, centosReleaseContent);
+            return Opt.stream(rhelProduct).flatMap(rhel -> {
+                if (!rhel.getSuseProduct().isPresent()) {
+                    LOG.warn("No product match found for: " + rhel.getName() + " " +
+                            rhel.getVersion() + " " + rhel.getRelease() + " " +
+                            server.getServerArch().getCompatibleChannelArch());
+                    return Stream.empty();
+                }
+                else {
+                    return Opt.stream(rhel.getSuseProduct());
+                }
+            }).collect(Collectors.toSet());
+        }
+        return Collections.emptySet();
+    }
+
+    public static ServerHistoryEvent addHistoryEvent(MinionServer server, String summary,
+            String details) {
+        ServerHistoryEvent historyEvent = new ServerHistoryEvent();
+        historyEvent.setCreated(new Date());
+        historyEvent.setServer(server);
+        historyEvent.setSummary(summary);
+        historyEvent.setDetails(details);
+        server.getHistory().add(historyEvent);
+        return historyEvent;
+    }
+
+    private static Stream<Channel> lookupRequiredChannelsForProduct(SUSEProduct sp) {
+        return RegistrationUtils.recommendedChannelsByBaseProduct(sp);
+    }
+
+    private static Stream<Channel> recommendedChannelsByBaseProduct(SUSEProduct base) {
+        return recommendedChannelsByBaseProduct(base, base);
+    }
+
+    private static Stream<Channel> recommendedChannelsByBaseProduct(SUSEProduct root, SUSEProduct base) {
+        return root.getSuseProductChannels().stream()
+                .filter(c -> c.getParentChannelLabel() == null)
+                .map(SUSEProductChannel::getChannelLabel)
+                .findFirst().map(rootChannelLabel -> {
+                    List<SUSEProduct> allExtensionProductsOf =
+                            SUSEProductFactory.findAllExtensionProductsOf(base);
+
+
+
+                    Stream<Channel> channelStream = SUSEProductFactory.findAllSUSEProductChannels().stream()
+                            .filter(pc -> pc.getProduct().equals(base))
+                            .map(SUSEProductChannel::getChannel)
+                            .filter(Objects::nonNull)
+                            .filter(c -> c.getParentChannel() == null ||
+                                    c.getParentChannel().getLabel().equals(rootChannelLabel));
+
+                    Stream<Channel> stream = allExtensionProductsOf.stream().flatMap(ext -> {
+                        return SUSEProductFactory.findSUSEProductExtension(root, base, ext).map(pe -> {
+                            if (pe.isRecommended()) {
+                                return recommendedChannelsByBaseProduct(root, ext);
+                            }
+                            else {
+                                return Stream.<Channel>empty();
+                            }
+                        }).orElseGet(Stream::empty);
+                    });
+
+                    return Stream.concat(
+                            channelStream,
+                            stream
+                    );
+                }).orElseGet(Stream::empty);
+    }
+
 }
